+++ evmd/tests/integration/x/feemarket/x_feemarket_test.go
@@ -1,4 +1,4 @@
-package integration
+package feemarket
 
 import (
 	"testing"
@@ -6,12 +6,13 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	evm "github.com/cosmos/evm"
+	"github.com/cosmos/evm/evmd/tests/integration"
 	"github.com/cosmos/evm/tests/integration/x/feemarket"
 	testapp "github.com/cosmos/evm/testutil/app"
 )
 
 func TestFeeMarketKeeperTestSuite(t *testing.T) {
-	create := testapp.ToEvmAppCreator[evm.IntegrationNetworkApp](CreateEvmd, "evm.IntegrationNetworkApp")
+	create := testapp.ToEvmAppCreator[evm.IntegrationNetworkApp](integration.CreateEvmd, "evm.IntegrationNetworkApp")
 	s := feemarket.NewTestKeeperTestSuite(create)
 	suite.Run(t, s)
 }
